+++ app/src/main/java/com/mybills/info/views/EditBillFragment.kt
@@ -103,7 +103,7 @@ class EditBillFragment : Fragment() ,DialogListener{
              && !TextUtils.isEmpty(ed_date.text.toString())){
             userViewModel.updateUser(User(uid,userName,ed_place_of_visit.text.toString(),
                 ed_purpose_visit.text.toString(),
-                ed_price.text.toString().toLong(), ed_date.text.toString()
+                ed_price.text.toString().toDouble(), ed_date.text.toString()
             ))
          }
      }
